user logout and fixing post deletion

+++ app/Http/Controllers/PostController.php
@@ -9,7 +9,7 @@ class PostController extends Controller {
 
 	public function getDashboard()
 	{
-		$posts = Post::all();
+		$posts = Post::orderBy('created_at','desc')->get();
 		return view('dashboard',['posts'=> $posts]);
 	}
 
@@ -31,6 +31,10 @@ class PostController extends Controller {
 	public function getDeletePost( $post_id ) 
 	{
 		$post = Post::where('id',$post_id)->first();
+		if(Auth::user() != $post->user)
+		{
+			return redirect()->back();
+		}
 		$post->delete();
 		return redirect()->route('dashboard')->with(['message' => 'Successfully Deleted !']);
 	}
